requirements.txt
beautifulsoup4==4.6.0
certifi==2019.9.11
chardet==3.0.4
dj-database-url==0.5.0
dj-static==0.0.6
Django==2.2.5
django-extensions==2.2.2
django-modelcluster==4.4
django-recaptcha==2.0.5
django-taggit==0.24.0
django-toolbelt==0.0.1
django-treebeard==4.3
djangorestframework==3.10.3
draftjs-exporter==2.1.6
gunicorn==19.9.0
html5lib==1.0.1
idna==2.8
Pillow==6.1.0
pkg-resources==0.0.0
psycopg2==2.8.3
psycopg2-binary==2.8.3
pytz==2019.2
requests==2.22.0
six==1.12.0
sqlparse==0.3.0
static3==0.7.0
Unidecode==1.1.1
urllib3==1.25.5
wagtail==2.6.2
wagtail-django-recaptcha==1.0
webencodings==0.5.1
Willow==1.1
